Splitting DMA header into multiple files
